+++ Assets/Scripts/Enemy/IDamageable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void SetDamage(float damage, Vector3 hitNormal);
+    void SetDamage(int damage, Vector3 hitNormal);
 }
